Hotfix: level calculation
