+++ app/src/main/java/com/itba/runningMate/levels/LevelsPresenter.kt
@@ -1,15 +1,12 @@
 package com.itba.runningMate.levels
 
 import com.itba.runningMate.domain.Level
-import com.itba.runningMate.repository.run.RunRepository
-import com.itba.runningMate.utils.providers.schedulers.SchedulerProvider
+import com.itba.runningMate.repository.aggregaterunmetrics.AggregateRunMetricsStorage
 import io.reactivex.disposables.CompositeDisposable
-import timber.log.Timber
 import java.lang.ref.WeakReference
 
 class LevelsPresenter(
-    private val repo: RunRepository,
-    private val schedulerProvider: SchedulerProvider,
+    private val aggregateRunMetricsStorage: AggregateRunMetricsStorage,
     view: LevelsView
 ) {
 
@@ -25,18 +22,9 @@ class LevelsPresenter(
     }
 
     private fun level() {
-        disposables.add(repo.getTotalDistance()
-            .subscribeOn(schedulerProvider.computation())
-            .observeOn(schedulerProvider.ui())
-            .subscribe({ distance: Double -> receivedTotalDistance(distance) }) { onReceivedTotalDistanceError() })
-    }
-
-    private fun receivedTotalDistance(distance: Double) {
+        val distance = aggregateRunMetricsStorage.getTotalDistance()
         val level = Level.from(distance)
         view.get()?.showCurrentLevel(level, distance)
     }
 
-    private fun onReceivedTotalDistanceError() {
-        Timber.d("Failed to retrieve total distance from db")
-    }
 }
