+++ app/src/main/java/com/itba/runningMate/mainpage/fragments/feed/FeedPresenter.kt
@@ -4,6 +4,7 @@ import com.itba.runningMate.domain.Achievements
 import com.itba.runningMate.domain.Level
 import com.itba.runningMate.domain.Run
 import com.itba.runningMate.repository.achievements.AchievementsRepository
+import com.itba.runningMate.repository.aggregaterunmetrics.AggregateRunMetricsStorage
 import com.itba.runningMate.repository.run.RunRepository
 import com.itba.runningMate.utils.providers.schedulers.SchedulerProvider
 import io.reactivex.disposables.CompositeDisposable
@@ -14,6 +15,7 @@ class FeedPresenter(
     private val repo: RunRepository,
     private val achievementsRepository: AchievementsRepository,
     private val schedulerProvider: SchedulerProvider,
+    private val aggregateRunMetricsStorage: AggregateRunMetricsStorage,
     view: FeedView
 ) {
 
@@ -71,10 +73,12 @@ class FeedPresenter(
     }
 
     private fun level() {
-        disposables.add(repo.getTotalDistance()
-            .subscribeOn(schedulerProvider.computation())
-            .observeOn(schedulerProvider.ui())
-            .subscribe({ distance: Double -> receivedTotalDistance(distance) }) { onReceivedTotalDistanceError() })
+        if (view.get() != null) {
+            view.get()?.stopLevelShimmerAnimation()
+            val distance = aggregateRunMetricsStorage.getTotalDistance()
+            val level = Level.from(distance)
+            view.get()!!.showCurrentLevel(level, distance)
+        }
     }
 
     private fun achievements() {
@@ -94,18 +98,6 @@ class FeedPresenter(
     private fun onReceivedAchievementsError() {
         Timber.d("Failed to retrieve completed achievements from db")
         view.get()?.showAchievements(listOf())
-    }
-
-    private fun receivedTotalDistance(distance: Double) {
-        if (view.get() != null) {
-            view.get()?.stopLevelShimmerAnimation()
-            val level = Level.from(distance)
-            view.get()!!.showCurrentLevel(level, distance)
-        }
-    }
-
-    private fun onReceivedTotalDistanceError() {
-        Timber.d("Failed to retrieve total distance from db")
     }
 
     fun goToAchievementsActivity() {
